divided edit post function
