+++ js/api/posts/editPost.js
@@ -1,42 +1,49 @@
-// import { updatePost } from "../../api.js";
+import { updatePost } from "./updatePost.js";
+import { getParams } from "../../utils/helpers/getParams.js";
+import { messageForUser } from "../../ui/messageForUser.js";
 
 /**
- * Edits an existing post.
+ * Edits a post.
+ *
+ * It takes an event object, prevents the default form submission behavior, and retrieves the form data.
+ * It then creates a post object from the form data and retrieves the post id from the URL parameters.
+ * If the id is not found, it throws an error.
+ * It then tries to update the post using the `updatePost` function.
+ * If the update is successful, it hides the form, clears the h1 element, displays a success message, and redirects to the profile page after 3 seconds.
+ * If the update fails, it displays an error message.
  *
  * @param {Event} event - The form submission event.
- * @throws {Error} If the post ID is not found in the URL.
+ * @throws {Error} If the post id is not found in the URL parameters.
  */
 
-// export async function editPost(event) {
-//   event.preventDefault();
+export async function editPost(event) {
+  event.preventDefault();
 
-//   const form = event.target;
-//   const formData = new FormData(form);
+  const form = event.target;
+  const formData = new FormData(form);
 
-//   let post = {};
-//   formData.forEach((value, key) => {
-//     post[key] = value;
-//   });
+  let post = {};
+  formData.forEach((value, key) => {
+    post[key] = value;
+  });
 
-//   const id = getParamFromUrl("id");
-//   if (!id) {
-//     throw new Error("Sorry, we couldn't find the post you're looking for.");
-//   }
+  const id = getParams("id");
+  if (!id) {
+    throw new Error("Sorry, we couldn't find the post you're looking for.");
+  }
 
-//   post.id = id;
+  console.log(post);
+  try {
+    await updatePost(post);
+    document.querySelector("form#editPostForm").style.display = "none";
+    document.querySelector("h1").innerHTML = "";
+    messageForUser("#messageForUser", "success", "Post updated successfully.");
 
-//   console.log(post);
-//   try {
-//     await updatePost(post);
-//     document.querySelector("form#editPostForm").style.display = "none";
-//     document.querySelector("h1").innerHTML = "";
-//     messageForUser("#messageForUser", "success", "Post updated successfully.");
-
-//     setTimeout(() => {
-//       window.location.href = "/profile/";
-//     }, 3000);
-//   } catch (error) {
-//     console.log(error);
-//     messageForUser("#messageForUser", "danger", "Sorry, we couldn't update the post.");
-//   }
-// }
+    setTimeout(() => {
+      window.location.href = "/profile/";
+    }, 3000);
+  } catch (error) {
+    console.log(error);
+    messageForUser("#messageForUser", "danger", "Sorry, we couldn't update the post.");
+  }
+}
