+++ js/handlers/posts/editPostHandler.js
@@ -1,68 +1,44 @@
 import { getSinglePost } from "../../api/posts/getSinglePost.js";
 import { messageForUser } from "../../ui/messageForUser.js";
-import { updatePost } from "../../api/posts/updatePost.js";
-import { getParamFromUrl } from "../../utils/helpers/getParams.js";
+import { getParams } from "../../utils/helpers/getParams.js";
+import { editPost } from "../../api/posts/editPost.js";
 
 /**
- * Handles the editing of a post.
+ * Handles the post editing process.
  *
- * It retrieves the post ID from the URL, fetches the post with that ID, and populates the edit form with the post's data.
- * If the fetch fails, the post ID is not found in the URL, or the form is not found, it displays an error message to the user.
+ * It retrieves the post id from the URL parameters. If the id is not found and the current URL path is the edit post page, it throws an error.
+ * If the id is found, it retrieves the post data using the `getSinglePost` function and fills the form with the post data.
+ * It also adds an event listener to the form to handle the form submission using the `editPost` function.
+ * If there's an error during this process, it displays an error message to the user.
+ *
+ * @throws {Error} If the post id is not found and the current URL path is the edit post page.
  */
 
 export async function editPostHandler() {
-  const id = getParamFromUrl("id");
+  const id = getParams("id");
 
-  if (!id) {
+  if (window.location.pathname === "/profile/edit-post.html" && !id) {
     throw new Error("Sorry, we couldn't find the post you're looking for.");
   }
 
-  try {
-    const post = await getSinglePost(id);
-    console.log(post);
-    const form = document.querySelector("#editPostForm");
-    form.id.value = post.id;
-    document.getElementById("formTitle").value = post.title;
-    document.getElementById("formMediaUrl").value = post.media;
-    document.getElementById("newPost").value = post.body;
-    form.addEventListener("submit", editPost);
-  } catch {
-    console.log(Error);
-    messageForUser("#messageForUser", "danger", "sorry, we couldn't load editing form.");
+  if (id) {
+    try {
+      const post = await getSinglePost(id);
+      populateForm(post);
+    } catch (error) {
+      console.log(Error);
+      messageForUser("#messageForUser", "danger", "sorry, we couldn't load editing form.");
+    }
   }
 }
 
 document.addEventListener("DOMContentLoaded", editPostHandler);
 
-//should i move this function out??????
-async function editPost(event) {
-  event.preventDefault();
-
-  const form = event.target;
-  const formData = new FormData(form);
-
-  let post = {};
-  formData.forEach((value, key) => {
-    post[key] = value;
-  });
-
-  const id = getParamFromUrl("id");
-  if (!id) {
-    throw new Error("Sorry, we couldn't find the post you're looking for.");
-  }
-
-  console.log(post);
-  try {
-    await updatePost(post);
-    document.querySelector("form#editPostForm").style.display = "none";
-    document.querySelector("h1").innerHTML = "";
-    messageForUser("#messageForUser", "success", "Post updated successfully.");
-
-    setTimeout(() => {
-      window.location.href = "/profile/";
-    }, 3000);
-  } catch (error) {
-    console.log(error);
-    messageForUser("#messageForUser", "danger", "Sorry, we couldn't update the post.");
-  }
+function populateForm(post) {
+  const form = document.querySelector("#editPostForm");
+  form.id.value = post.id;
+  document.getElementById("formTitle").value = post.title;
+  document.getElementById("formMediaUrl").value = post.media;
+  document.getElementById("newPost").value = post.body;
+  form.addEventListener("submit", editPost);
 }
