+++ js/ui/displayUserName.js
@@ -7,14 +7,15 @@ import { getUserName } from "../utils/helpers/getUserName.js";
  * If the user's name is not found, it logs a message to the console.
  */
 
-document.addEventListener("DOMContentLoaded", () => {
-  displayUserName();
-});
-
 export function displayUserName() {
   const userName = getUserName();
   if (userName) {
-    document.querySelector(".your-profile-name").textContent = userName;
+    const userNameElement = document.querySelector(".your-profile-name");
+    if (userNameElement) {
+      userNameElement.textContent = userName;
+    } else {
+      console.error("Element with class .your-profile-name not found");
+    }
   } else {
     console.log("User name not found in storage.");
   }
