+++ profile/edit-post.html
@@ -8,8 +8,8 @@
       content="Vib'n: A cool place for cool people!. Dive into a world where vibes connect, friendships form, and fun never ends. Join the community that's redefining social interaction!"
     />
     <title>Vib`n | Feed | Edit Post</title>
-    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css" />
     <link rel="stylesheet" href="/dist/css/style.css" />
+    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css" />
     <link rel="stylesheet" href="4circle-line-bounce/style.css" />
     <script
       defer
