+++ profile/index.html
@@ -8,8 +8,8 @@
       content="Vib'n: A cool place for cool people!. Dive into a world where vibes connect, friendships form, and fun never ends. Join the community that's redefining social interaction!"
     />
     <title>Vib`n | Profile</title>
-    <!-- <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css" /> -->
     <link rel="stylesheet" href="/dist/css/style.css" />
+    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css" />
     <script
       defer
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
